requirements.txt
dash>=2.14.0
plotly>=5.17.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

